Refactoring the controller and changing facebook route
